Refactor of selectMenuHandler

+++ src/handlers/selectMenuHandler.js
@@ -5,36 +5,29 @@ module.exports = async (client, interaction) => {
     let instance = client.readInstanceFile(guildId);
     let rustplus = client.rustplusInstances[guildId];
 
-    let command = (interaction.customId.endsWith('AutoDayNight')) ? 'AutoDayNight' : 'prefix';
-    let id = (command === 'AutoDayNight') ? interaction.customId.replace('AutoDayNight', '') : 0;
+    if (interaction.customId === 'prefix') {
+        instance.generalSettings.prefix = interaction.values[0];
+        client.writeInstanceFile(guildId, instance);
 
-    switch (command) {
-        case 'prefix':
-            instance.generalSettings.prefix = interaction.values[0];
-            client.writeInstanceFile(guildId, instance);
+        if (rustplus) {
+            rustplus.generalSettings.prefix = interaction.values[0];
+        }
 
-            if (rustplus) {
-                rustplus.generalSettings.prefix = interaction.values[0];
-            }
+        let row = DiscordTools.getPrefixSelectMenu(interaction.values[0]);
+        await interaction.update({ components: [row] });
+    }
+    else if (interaction.customId.endsWith('AutoDayNight')) {
+        let id = interaction.customId.replace('AutoDayNight', '');
 
-            let row = DiscordTools.getPrefixSelectMenu(interaction.values[0]);
-            await interaction.update({ components: [row] });
-            break;
+        if (!instance.switches.hasOwnProperty(id)) {
+            await client.switchesMessages[guildId][id].delete();
+            delete client.switchesMessages[guildId][id];
+            return;
+        }
 
-        case 'AutoDayNight':
-            if (!instance.switches.hasOwnProperty(id)) {
-                await client.switchesMessages[guildId][id].delete();
-                delete client.switchesMessages[guildId][id];
-                return;
-            }
+        instance.switches[id].autoDayNight = parseInt(interaction.values[0]);
+        client.writeInstanceFile(guildId, instance);
 
-            instance.switches[id].autoDayNight = parseInt(interaction.values[0]);
-            client.writeInstanceFile(guildId, instance);
-
-            DiscordTools.sendSmartSwitchMessage(guildId, id, false, true, false, interaction);
-            break;
-
-        default:
-            break;
+        DiscordTools.sendSmartSwitchMessage(guildId, id, false, true, false, interaction);
     }
 }
